place order temp

+++ TradingBot/OrderDetails.cs
@@ -9,8 +9,8 @@ namespace TradingBot
     }
     public struct OrderDetails
     {
-        public float limitPrice;
-        public float quantity;
+        public decimal limitPrice;
+        public int quantity;
         public string symbol;
         public TimeInForce timeInForce;
         public OrderType orderType;
